Update to V1.21.3; Apply patch to force no URI lookup
The patch to force no URI lookup is supposed to fix an issue that occurs when using the .local TLD which clashes with Bonjour/mDNS. If URI lookup is enabled, DNS URI resolve requests will be done via mDNS which takes a long time and eventually fails. Afterwards, the SRV record is queried as a fallback which works fine. Since Windows DNS server does not even support URI records at the moment it should ok to disable it by default at compile time. Users can still re-enable it in their krb5.conf if desired.